[server] Do synchronous I/O multiplexing for game threads
The server can listen to both clients, instead of just the one who has
the turn, as it used to be the case.

Based on "Beej’s Guide to Network Programming", section 7.2
"`poll()`---Synchronous I/O Multiplexing":

> "So how can you avoid polling? Not slightly ironically, you can avoid
polling by using the `poll()` system call. In a nutshell, we’re going to
ask the operating system to do all the dirty work for us, and just let us
know when some data is ready to read on which sockets. In the meantime,
our process can go to sleep, saving system resources."