[client+server] Generalise and handle `END_GAME` for the three reasons it can occur
This makes it easier to parse for the client, as they have a predictable
structure: END_GAME <reason> <player>.

I had some guidance from GPT 4o to use a static lookup table for the
enum, which I had already read about and used, but had forgotten the
technique since. It might not be, admittedly, the most legible code, but
it handles all the possible cases in one place, which is desirable.