[server] Fix `accept` return value comparison
I unintentionally introduced this error when removing the `!= -1` from
comparisons as I thought that looked a bit "ugly" in commit
69c76e7dfd44.

I opted for putting the constant on the left hand side of the comparison
as suggested by this C Coding Standard by Eno Thereska at Carnegie
Mellon:
https://users.ece.cmu.edu/~eno/coding/CCodingStandard.html#ifthen

> One reason is that if you leave out one of the = signs, the compiler
will find the error for you. A second reason is that it puts the value
you are looking for right up front where you can find it instead of
buried at the end of your expression. It takes a little time to get used
to this format, but then it really gets useful.